Major module restructuring for publication
