+++ docs/api.rst
@@ -1,0 +1,24 @@
+The cellocity API reference
+===========================
+
+.. automodule:: cellocity
+   :members:
+   
+The "channel" module
+--------------------
+
+.. automodule:: cellocity.channel
+   :members:
+   
+The "analysis" module
+---------------------
+
+.. automodule:: cellocity.analysis
+   :memebers:
+   
+The "validation" module
+-----------------------
+
+.. automodule:: cellocity.validation
+   :members:
+
