+++ docs/developer.rst
@@ -1,0 +1,5 @@
+Developer Information
+=====================
+
+Develop responsibly
+
