+++ docs/tutorial.rst
@@ -1,0 +1,13 @@
+Cellocity Tutorial
+==================
+
+Step-by-step guide
+------------------
+
+Do like this:
+
+1. load
+2. Process
+3. Analyse data
+
+
